--- model/user_model.ts
@@ -1,8 +1,0 @@
-'use strict';
-
-import {buildUser} from "./user";
-import {userValidator} from "./user_validator";
-import {userSchema} from "./user_schema";
-
-const validator = userValidator(userSchema);
-export const makeUser = buildUser(validator);
