+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "server-base-serverlss",
+  "name": "server-base-serverless",
   "version": "0.0.1",
   "description": "Moboweb Serverless boilerplate",
   "main": "index.js",
